+++ seantis/dir/events/tests/test_eventindex.py
@@ -1,13 +1,14 @@
 import transaction
 
-from datetime import date
+from datetime import date, datetime
+from seantis.dir.events import dates
 from seantis.dir.events.tests import IntegrationTestCase
 
 from seantis.dir.events.catalog import (
     LazyList,
     EventOrderIndex,
     attach_reindex_to_transaction,
-    ReindexDataManager
+    ReindexDataManager,
 )
 
 from mock import Mock
@@ -387,3 +388,29 @@ class TestEventIndex(IntegrationTestCase):
         self.assertEqual(1, number_of_data_managers())
         attach_reindex_to_transaction(self.directory)
         self.assertEqual(1, number_of_data_managers())
+
+    def test_by_range_timezone_regression(self):
+        self.login_testuser()
+
+        published = self.catalog.indices['published']
+
+        self.assertEqual(len(published.index), 0)
+
+        event = self.create_event(
+            start=datetime(2014, 12, 31, 22),
+            end=datetime(2014, 12, 31, 23),
+            timezone='Europe/Vienna'
+        )
+        event.submit()
+        event.publish()
+        transaction.commit()
+
+        self.assertEqual(len(published.index), 1)
+
+        dtrange = (datetime(2015, 1, 1), datetime(2015, 12, 31))
+        dtrange = [dates.as_timezone(dt, 'Europe/Vienna') for dt in dtrange]
+        self.assertEqual(len(published.by_range(*dtrange)), 0)
+
+        dtrange = (datetime(2014, 1, 1), datetime(2014, 12, 31, 23, 59))
+        dtrange = [dates.as_timezone(dt, 'Europe/Vienna') for dt in dtrange]
+        self.assertEqual(len(published.by_range(*dtrange)), 1)