+++ docs/HISTORY.rst
@@ -2,8 +2,12 @@
 Changelog
 ---------
 
-1.1rc1 (unreleased)
-~~~~~~~~~~~~~~~~~~~
+1.0.1 (unreleased)
+~~~~~~~~~~~~~~~~~~
+
+- Fixes a rare bug occurring when an event exist for the last day of a year,
+  but no events exist for the whole next year. Fixes #37.
+  [href]
 
 
 1.0